+++ ocl_kernels/kernel.cpp
@@ -1,0 +1,10 @@
+#include "kernel.hpp"
+#include <typeinfo>
+#include <unistd.h>
+#include <string.h>
+#include <errno.h>
+
+
+void Kernel::init(int n) {
+
+}
